requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues


azure-functions
azure-cosmos>=4.0.0
azure-storage-blob>=12.0.0
azure-identity>=1.8.0
